+++ client/src/views/Firebase/GoogleSign.jsx
@@ -1,4 +1,3 @@
-import "../register/register.css";
 import { useDispatch } from "react-redux";
 import { StartGoogleAuth } from "../../redux/apiPetitions";
